+++ oci-runtime-spec/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "oci_runtime_spec"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0.126", features = ["derive"] }
+
+[dev-dependencies]
+serde_json = "1.0.64"
