+++ sprofiler-bpf/Cargo.toml
@@ -18,3 +18,4 @@ sprofiler-sys = { path="../sprofiler-sys" }
 serde_json = "1.0.62"
 nix = "0.20.0"
 signal-hook = { version = "0.3.6", features = ["extended-siginfo", "iterator"]  }
+oci_runtime_spec = { path = "../oci-runtime-spec" }
