+++ sprofiler-sys/Cargo.toml
@@ -15,3 +15,4 @@ regex = "1.4.3"
 serde = "1.0.123"
 serde_derive = "1.0.121"
 serde_json = "1.0.62"
+oci_runtime_spec = { path = "../oci-runtime-spec" }
